fix: Unified API code review (#144)
