+++ dart/lib/src/transport/noop_transport.dart
@@ -5,8 +5,5 @@ import 'transport.dart';
 
 class NoOpTransport implements Transport {
   @override
-  Dsn get dsn => null;
-
-  @override
   Future<SentryId> send(SentryEvent event) => Future.value(SentryId.empty());
 }
