+++ dart/test/sentry_event_test.dart
