+++ dart/test/test_utils.dart
@@ -8,6 +8,7 @@ import 'dart:convert';
 import 'package:http/http.dart';
 import 'package:http/testing.dart';
 import 'package:sentry/sentry.dart';
+import 'package:sentry/src/transport/http_transport.dart';
 import 'package:test/test.dart';
 
 const String testDsn = 'https://public:secret@sentry.example.com/1';
@@ -88,7 +89,8 @@ Future testCaptureException(
     expect('$sentryId', 'testeventid');
   }
 
-  expect(postUri, options.transport.dsn.postUri);
+  final transport = options.transport as HttpTransport;
+  expect(postUri, transport.dsn.postUri);
 
   testHeaders(
     headers,
@@ -178,55 +180,67 @@ void runTest({Codec<List<int>, List<int>> gzip, bool isWeb = false}) {
   test('can parse DSN', () async {
     final options = SentryOptions(dsn: testDsn);
     final client = SentryClient(options);
-    expect(options.transport.dsn.uri, Uri.parse(testDsn));
+
+    final transport = options.transport as HttpTransport;
+
+    expect(transport.dsn.uri, Uri.parse(testDsn));
     expect(
-      options.transport.dsn.postUri,
+      transport.dsn.postUri,
       'https://sentry.example.com/api/1/store/',
     );
-    expect(options.transport.dsn.publicKey, 'public');
-    expect(options.transport.dsn.secretKey, 'secret');
-    expect(options.transport.dsn.projectId, '1');
+    expect(transport.dsn.publicKey, 'public');
+    expect(transport.dsn.secretKey, 'secret');
+    expect(transport.dsn.projectId, '1');
     await client.close();
   });
 
   test('can parse DSN without secret', () async {
     final options = SentryOptions(dsn: _testDsnWithoutSecret);
     final client = SentryClient(options);
-    expect(options.transport.dsn.uri, Uri.parse(_testDsnWithoutSecret));
+
+    final transport = options.transport as HttpTransport;
+
+    expect(transport.dsn.uri, Uri.parse(_testDsnWithoutSecret));
     expect(
-      options.transport.dsn.postUri,
+      transport.dsn.postUri,
       'https://sentry.example.com/api/1/store/',
     );
-    expect(options.transport.dsn.publicKey, 'public');
-    expect(options.transport.dsn.secretKey, null);
-    expect(options.transport.dsn.projectId, '1');
+    expect(transport.dsn.publicKey, 'public');
+    expect(transport.dsn.secretKey, null);
+    expect(transport.dsn.projectId, '1');
     await client.close();
   });
 
   test('can parse DSN with path', () async {
     final options = SentryOptions(dsn: _testDsnWithPath);
     final client = SentryClient(options);
-    expect(options.transport.dsn.uri, Uri.parse(_testDsnWithPath));
+
+    final transport = options.transport as HttpTransport;
+
+    expect(transport.dsn.uri, Uri.parse(_testDsnWithPath));
     expect(
-      options.transport.dsn.postUri,
+      transport.dsn.postUri,
       'https://sentry.example.com/path/api/1/store/',
     );
-    expect(options.transport.dsn.publicKey, 'public');
-    expect(options.transport.dsn.secretKey, 'secret');
-    expect(options.transport.dsn.projectId, '1');
+    expect(transport.dsn.publicKey, 'public');
+    expect(transport.dsn.secretKey, 'secret');
+    expect(transport.dsn.projectId, '1');
     await client.close();
   });
   test('can parse DSN with port', () async {
     final options = SentryOptions(dsn: _testDsnWithPort);
     final client = SentryClient(options);
-    expect(options.transport.dsn.uri, Uri.parse(_testDsnWithPort));
+
+    final transport = options.transport as HttpTransport;
+
+    expect(transport.dsn.uri, Uri.parse(_testDsnWithPort));
     expect(
-      options.transport.dsn.postUri,
+      transport.dsn.postUri,
       'https://sentry.example.com:8888/api/1/store/',
     );
-    expect(options.transport.dsn.publicKey, 'public');
-    expect(options.transport.dsn.secretKey, 'secret');
-    expect(options.transport.dsn.projectId, '1');
+    expect(transport.dsn.publicKey, 'public');
+    expect(transport.dsn.secretKey, 'secret');
+    expect(transport.dsn.projectId, '1');
     await client.close();
   });
   test('sends client auth header without secret', () async {
